server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
openai>=1.52.0
requests>=2.32.0
python-dotenv>=1.0.1
duckdb>=1.1.0
